Restructure folders for clarity
