--- OuladEtlEda/BulkLoader.cs
@@ -1,12 +1,0 @@
-using EFCore.BulkExtensions;
-using Microsoft.EntityFrameworkCore;
-
-namespace OuladEtlEda;
-
-public class BulkLoader
-{
-    public async Task BulkInsertAsync<T>(DbContext context, IList<T> entities) where T : class
-    {
-        await context.BulkInsertAsync(entities);
-    }
-}
