--- OuladEtlEda/Csv/AssessmentCsv.cs
@@ -1,18 +1,0 @@
-using CsvHelper.Configuration.Attributes;
-
-namespace OuladEtlEda.Csv;
-
-public class AssessmentCsv
-{
-    [Name("id_assessment")] public int IdAssessment { get; set; }
-
-    [Name("code_module")] public string CodeModule { get; set; } = null!;
-
-    [Name("code_presentation")] public string CodePresentation { get; set; } = null!;
-
-    [Name("assessment_type")] public string? AssessmentType { get; set; }
-
-    [Name("date")] public int? Date { get; set; }
-
-    [Name("weight")] public int Weight { get; set; }
-}
