--- OuladEtlEda/DomainException.cs
@@ -1,16 +1,0 @@
-namespace OuladEtlEda;
-
-public class DomainException : Exception
-{
-    public DomainException()
-    {
-    }
-
-    public DomainException(string message) : base(message)
-    {
-    }
-
-    public DomainException(string message, Exception inner) : base(message, inner)
-    {
-    }
-}
