--- OuladEtlEda/IDomainValidator.cs
@@ -1,6 +1,0 @@
-namespace OuladEtlEda;
-
-public interface IDomainValidator<T>
-{
-    Task ValidateAsync(T entity);
-}
