+++ src/Domain/DomainException.cs
@@ -1,0 +1,16 @@
+namespace OuladEtlEda.Domain;
+
+public class DomainException : Exception
+{
+    public DomainException()
+    {
+    }
+
+    public DomainException(string message) : base(message)
+    {
+    }
+
+    public DomainException(string message, Exception inner) : base(message, inner)
+    {
+    }
+}
