+++ src/Domain/IDomainValidator.cs
@@ -1,0 +1,6 @@
+namespace OuladEtlEda.Domain;
+
+public interface IDomainValidator<T>
+{
+    Task ValidateAsync(T entity);
+}
