+++ src/Pipeline/EtlException.cs
@@ -1,0 +1,16 @@
+namespace OuladEtlEda.Pipeline;
+
+public class EtlException : Exception
+{
+    public EtlException()
+    {
+    }
+
+    public EtlException(string message) : base(message)
+    {
+    }
+
+    public EtlException(string message, Exception inner) : base(message, inner)
+    {
+    }
+}
