+++ tests/MapperTests/CategoricalOrdinalMapperTests.cs
